+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.0.3] - 2022-xx-xx
 
+### Added
+
+- added onRef to work outside React
+
 ### Changed
 
 - better onChange